requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.5.3
Pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
